change splash activity to splash fragment
