--- app/src/main/java/com/epicteam1/skimountains/SplashActivity.kt
@@ -1,33 +1,0 @@
-package com.epicteam1.skimountains
-
-import android.annotation.SuppressLint
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.epicteam1.skimountains.feature_ski_places.presentation.viewModel.SkiViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-
-@SuppressLint("CustomSplashScreen")
-class SplashActivity : AppCompatActivity() {
-
-    companion object {
-        var startCount = 0
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-
-        supportActionBar?.hide()
-
-        CoroutineScope(Dispatchers.Main).launch {
-            delay(1600L)
-            startCount++
-            startActivity(Intent(this@SplashActivity, MainActivity::class.java))
-            finish()
-        }
-    }
-}
