+++ app/src/main/java/com/epicteam1/skimountains/feature_ski_places/presentation/ui/SplashFragment.kt
@@ -1,0 +1,28 @@
+package com.epicteam1.skimountains.feature_ski_places.presentation.ui
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+import com.epicteam1.skimountains.R
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+
+class SplashFragment : Fragment(R.layout.fragment_splash) {
+
+    companion object {
+        var startCount = 0
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        CoroutineScope(Dispatchers.Main).launch {
+            delay(1600L)
+            startCount++
+            findNavController().navigate(R.id.action_splash_to_home_fragment)
+        }
+    }
+
+}
